Proyecto/backend_flask/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Session==0.5.0
psycopg2-binary==2.9.11
python-dotenv==1.0.0
bcrypt==4.0.1
docxtpl==0.16.7
docx2pdf==0.1.8
pywin32>=307
xhtml2pdf==0.2.17
pandas>=2.1.3
openpyxl==3.1.2
